refactor(wrapped): extract share flow into useWrappedshare — wrapped 100%
Move the share-token state, lazy mint (handleOpenShare), shareUrl, and
copy/native-share handlers verbatim into _hooks/use-wrapped-share.ts
(owns its own Wrapped translations). Pruned now-unused toast/supabase/
nativeShareOrCopy/useEffect imports. wrapped/page is now a lean
orchestrator: useWrappedData + useWrappedShare + useStoryPlayback +
render. Behavior-identical; wrapped/page 438->371 LOC (1493->371, -75%
across the sweep).
